Alterada a consulta de variacoes do Produto para o Servico Publico
