+++ src/Announcement.php
@@ -10,8 +10,6 @@ namespace Dsc\MercadoLivre;
 use Dsc\MercadoLivre\Announcement\Announcement as AnnouncementInterface;
 use Dsc\MercadoLivre\Announcement\AnnouncementManager;
 use Dsc\MercadoLivre\Announcement\ItemResponse;
-use Dsc\MercadoLivre\Announcement\Variation;
-use Doctrine\Common\Collections\Collection;
 
 class Announcement extends BaseService implements AnnouncementManager
 {
@@ -84,33 +82,6 @@ class Announcement extends BaseService implements AnnouncementManager
         return $this->getResponse(
             $this->put('/items/' . $code . '/description', $data),
             ItemResponse::class
-        );
-    }
-
-    /**
-     * @param string $code
-     * @return Collection
-     * @link https://developers.mercadolibre.com/pt_br/variacoes#Consultar-varia%C3%A7%C3%B5es
-     */
-    public function findVariations($code)
-    {
-        return $this->getResponse(
-            $this->get("/items/$code/variations"),
-            Variation::class
-        );
-    }
-
-    /**
-     * @param string $code - codigo do Produto Mercado Livre
-     * @param integer $id - ID da Variacao
-     * @return Variation
-     * @link https://developers.mercadolibre.com/pt_br/variacoes#Consultar-varia%C3%A7%C3%B5es
-     */
-    public function getVariation($code, $id)
-    {
-        return $this->getResponse(
-            $this->get("/items/$code/variations/$id"),
-            Variation::class
         );
     }
 }
